refac: Overall tour restructuring
-Deleted tour tourist view
-Merged tourist view into tour authoring
-Renamed tour-author-view to tour-detailed-view
