+++ Explorer/src/app/feature-modules/layout/landing/landing.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { TourCardService } from '../../tour-tourist/tour-card.service';
-import { TourCard } from '../../tour-tourist/model/tour-card.model';
+import { TourCard } from '../../tour-authoring/model/tour-card.model';
+import { TourAuthoringService } from '../../tour-authoring/tour-authoring.service';
 //import { Blog } from '../../blog/model/blog.model';
 
 @Component({
@@ -12,12 +12,12 @@ export class LandingComponent {
   recommendedTours: TourCard[] = [];
   //recommendedBlogs: Blog[] = [];
 
-  constructor(private tourCardService: TourCardService){
+  constructor(private tourService: TourAuthoringService){
     this.loadTours();
   }
 
   loadTours(): void{
-    this.tourCardService.getTourCards(1,4).subscribe({
+    this.tourService.getPublishedTourCards(1,4).subscribe({
       next: (result: TourCard[]) => {
         this.recommendedTours = result
         console.log(this.recommendedTours)
