+++ Explorer/src/app/feature-modules/layout/layout.module.ts
@@ -10,7 +10,6 @@ import { CardLinkComponent } from './card-link/card-link.component';
 import { FooterComponent } from './footer/footer.component';
 import { TourAuthoringModule } from "../tour-authoring/tour-authoring.module";
 import { SharedModule } from 'src/app/shared/shared.module';
-import { TourTouristModule } from '../tour-tourist/tour-tourist.module';
 
 @NgModule({
   declarations: [
@@ -26,7 +25,6 @@ import { TourTouristModule } from '../tour-tourist/tour-tourist.module';
     RouterModule,
     TourAuthoringModule,
     SharedModule,
-    TourTouristModule
 ],
   exports: [
     NavbarComponent,
