+++ Explorer/src/app/feature-modules/tour-authoring/tour-authoring.module.ts
@@ -11,9 +11,10 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { TourFormComponent } from './tour-form/tour-form.component';
 import { TourCreationComponent } from './tour-creation/tour-creation.component';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { TourMapComponent } from './tour-map/tour-map.component';
 import { SharedModule } from 'src/app/shared/shared.module';
-import { TourAuthorViewComponent } from './tour-view/tour-view.component';
+import { TourDetailedViewComponent } from './tour-view/tour-view.component';
+import { ToursPageComponent } from './tours-page/tours-page.component';
+import { TourCardComponent } from './tour-card/tour-card.component';
 
 @NgModule({
   declarations: [
@@ -24,9 +25,10 @@ import { TourAuthorViewComponent } from './tour-view/tour-view.component';
     FacilityDialogComponent,
     TourFormComponent,
     TourCreationComponent,
-    TourAuthorViewComponent,
+    TourDetailedViewComponent,
     TourEquipmentDialogComponent,
-    TourMapComponent
+    ToursPageComponent,
+    TourCardComponent
   ],
   imports: [
     CommonModule,
@@ -39,6 +41,7 @@ import { TourAuthorViewComponent } from './tour-view/tour-view.component';
     KeyPointsComponent,
     KeyPointFormComponent,
     TourComponent,
+    TourCardComponent
   ]
 })
 export class TourAuthoringModule { }
