+++ Explorer/src/app/feature-modules/tour-authoring/tour-card/tour-card.component.css
