+++ Explorer/src/app/feature-modules/tour-authoring/tours-page/tours-page.component.css
