--- Explorer/src/app/feature-modules/tour-tourist/tour-tourist.module.ts
@@ -1,19 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { TourCardComponent } from './tour-card/tour-card.component';
-import { ToursPageComponent } from './tours-page/tours-page.component';
-
-
-@NgModule({
-  declarations: [
-    TourCardComponent,
-    ToursPageComponent
-  ],
-  imports: [
-    CommonModule
-  ],
-  exports:[
-    TourCardComponent,
-  ]
-})
-export class TourTouristModule { }
